fix(vbc): TIER-DETECT wire code follows the stdlib contract — AOT answers 3, not 1
core/intrinsics/runtime/tier.vr documents get_tier() as 0=interpreter,
1/2=JIT tiers, 3=AOT. The first lowering emitted 1 (the execution-
model 'Tier 1 = Aot' naming, not the intrinsic's wire contract), which
the tier suite's canonical-code pins caught at AOT (t==0||t==3). MLIR
GPU path and the sub-op doc updated to the same authority. Tier suite
13/13 both tiers.

+++ crates/verum_vbc/src/intrinsics/lowering.rs
@@ -1332,12 +1332,14 @@ impl IntrinsicLowering {
 
             // Tier detection (TIER-DETECT-AOT-1): GPU kernels are
             // compiled ahead-of-time — the honest per-tier answer on
-            // this path is "not the interpreter" (tier 1 shape).
+            // this path is "not the interpreter". Wire code 3 = AOT
+            // per the stdlib contract (tier.vr: 0=interp, 1/2=JIT,
+            // 3=AOT).
             InlineSequenceId::ExecutionTier => self.emit(MlirOp {
                 name: "arith.constant".to_string(),
                 attrs: vec![MlirAttr {
                     name: "value".to_string(),
-                    value: MlirAttrValue::Integer(1),
+                    value: MlirAttrValue::Integer(3),
                 }],
                 result_types: vec![MlirType::I64],
                 operands: vec![],
